SERVER-52587 Making collection and database instances comparable through cluster time (no more epochs)
Part 1: setting up the environment to start comparing DatabaseVersions
through timestamps.

Changelog:

- Moving the ComparableDatabaseVersion class to the DatabaseVersion
  header.
- Using ComparableDatabaseVersions in two places where we were doing
 DatabaseVersion comparsions.
